Nightly updated

+++ hanga/deploy_stack.py
@@ -17,29 +17,48 @@ from . import hanga_util as util
 from . import upload_object
 
 def deploy_stack(name, template, bucket, object_prefix, params, tags, upload, iam, named_iam, auto_expand, default, isUpdated):
+    if not default and template is None:
+        click.secho('Either --template (-t) or --default (-d) parameter is required.', bg=const.BG_ERROR, fg=const.FG_ERROR)
+        sys.exit(const.ERC_OTHERS)   
+
+    if default and (template is not None or params is not None or tags is not None):
+        click.secho('The --default (-d) parameter cannot be used with template, params, or tags pameter.', bg=const.BG_ERROR, fg=const.FG_ERROR)
+        sys.exit(const.ERC_OTHERS)     
+
+    if default:
+        template = name + const.YAML_EXT
+        params = name + '-params' + const.JSON_EXT
+        tags = name + '-tags' + const.JSON_EXT    
+        click.secho('Use the default file option.', fg=const.FG_INF)
+
     object_prefix = util.reformS3Prefix(object_prefix)
+
     try:
         with open(template, 'r') as fTemplate:
-            object_key = object_prefix + os.path.basename(fTemplate.name)
+            baseName = os.path.basename(fTemplate.name)
+            object_key = object_prefix + baseName
+            click.secho('Use the template file: %s' % baseName, fg=const.FG_INF)
     except FileNotFoundError:
         util.handleFileNotFound(template)
 
     if params is not None:
-        paramList = util.getJsonDataFromFile(params)
-        click.secho('Use the parameter file', fg=const.FG_INF)
+        baseName = os.path.basename(params)        
+        paramList = util.getJsonDataFromFile(params, baseName)
+        click.secho('Use the parameter file: %s' % baseName, fg=const.FG_INF)
     else:
         paramList = list()
          
     if tags is not None:
-        tagList = util.getJsonDataFromFile(tags)
-        click.secho('Use the tag file', fg=const.FG_INF) 
+        baseName = os.path.basename(tags)          
+        tagList = util.getJsonDataFromFile(tags, baseName)
+        click.secho('Use the tag file: %s' % baseName, fg=const.FG_INF) 
     else:
         tagList = list()           
 
     if (upload):
         click.secho('The template is being uploaded to the bucket.', fg=const.FG_INF)
         upload_object._upload_object(bucket, object_key, os.path.abspath(template))
-        click.secho('Upload done...', fg=const.FG_INF)
+        click.secho('The template has been uploaded to the bucket.', fg=const.FG_INF)
     
     templateUrl = 'https://' + bucket + '.s3.amazonaws.com/' + object_key
     try:
@@ -59,20 +78,33 @@ def deploy_stack(name, template, bucket, object_prefix, params, tags, upload, ia
         click.secho('and deployed to the stack:', fg=const.FG_INF)
         click.echo(stackId)  
 
-        response = _wait_for_done_changeset(name, cName)
-        click.echo(response)
-        
-        # TODO: Change set results to be shown for confirmation
-        
-        isYes = util.query_yes_no('\nDo you want to execute this change set?', 'no')
-        if not isYes:
-            _session.cf.delete_stack(StackName=name)
-            click.secho('You have aborted this change set.', fg=const.FG_INF) 
-            sys.exit(0)
+        response = _wait_for_created_changeset(name, cName)
+        # click.echo(response)
+        csExecutionStatus = response[const.CS_EXECUTION_STATUS]
 
-        response = _session.cf.execute_change_set(StackName=name,
-                                            ChangeSetName=cName)  
-        click.secho('The change set is being deployed.', fg=const.FG_INF)                                                     
+        if csExecutionStatus == const.CS_AVAILABLE:
+            click.secho('The change set is ready!', fg=const.FG_INF)
+
+            # TODO: Change set results to be shown for confirmation
+            _describe_changeset(name, cName, response)
+
+            isYes = util.query_yes_no('\nDo you want to execute this change set?', 'no')
+            if not isYes:
+                if isUpdated:
+                    _session.cf.delete_change_set(StackName=name,
+                                                    ChangeSetName=cName)
+                else:
+                    _session.cf.delete_stack(StackName=name)
+                click.secho('You have aborted this change set.', fg=const.FG_INF) 
+            else:
+                response = _session.cf.execute_change_set(StackName=name,
+                                                    ChangeSetName=cName)  
+                click.secho('The change set is being deployed.', fg=const.FG_INF)    
+                response = _wait_for_executed_changeset(name)  
+                click.secho('\nThe executed change set is done!', fg=const.FG_INF)
+        else:
+            click.secho('Nothing to be created/updated.', fg=const.FG_INF) 
+
     except botocore.exceptions.ClientError as error:
         util.handleClientError(error)
     # except:
@@ -80,13 +112,48 @@ def deploy_stack(name, template, bucket, object_prefix, params, tags, upload, ia
     #     sys.exit(const.ERC_OTHERS)  
 
 
-# def _describe_change_set(StackName=name, ChangeSetName=cName):
-#     # NextToken to be included
-#     response = session.cf.client.describe_change_set(StackName=name,
-#                                         ChangeSetName=cName)
-#     return response
+def _describe_changeset(name, cName, response):
+    dictChanges = {const.CS_ADD: [],
+                    const.CS_REMOVE: [], 
+                    const.CS_MODIFY: []}
+    
+    while True:
+        changes = response[const.CS_CHANGES]
+        for change in changes:
+            rChange = change[const.CS_RESOURCE_CHANGE]
+            action = rChange[const.CS_ACTION]
+            resourceType = rChange[const.CS_RESOURCE_TYPE]
+            logicalResourceId = rChange[const.CS_LOGICAL_RESOURCE_ID]
+            replacement = rChange.get(const.CS_REPLACEMENT)  
+            replacement = replacement if replacement is not None else 'False'
+            dictChanges[action].append((resourceType, logicalResourceId, replacement))
 
-def _wait_for_done_changeset(name, cName):   
+        nextToken = response.get(const.NEXT_TOKEN)
+        if nextToken is None:
+            break
+        else:
+            response = _session.cf.describe_change_set(StackName=name,
+                                                        ChangeSetName=cName,
+                                                        NextToken=nextToken) 
+    
+    # Summary of resources
+    lActions = [const.CS_ADD, const.CS_MODIFY, const.CS_REMOVE]
+    pActions = ['added', 'modified', 'removed']
+    sActions = ['+', '*', '-']
+    for i in range(3):
+        if len(dictChanges[lActions[i]]) > 0:
+            sAction = sActions[i]
+            click.secho('\nThe following resources will be %s:' % pActions[i], fg=const.FG_INF) 
+            for resource in dictChanges[lActions[i]]:
+                resourceType, logicalResourceId, replacement = resource
+                if replacement == 'False':
+                    click.echo(' [%s] %s having ID: %s.' % (sAction, resourceType, logicalResourceId))
+                elif replacement == 'True':
+                    click.echo(' [%s] %s having ID: %s to be replaced.' % (sAction, resourceType, logicalResourceId))
+                else:
+                    click.echo(' [%s] %s having ID: %s to be conditionally replaced.' % (sAction, resourceType, logicalResourceId))
+
+def _wait_for_created_changeset(name, cName):   
     click.secho('Please wait while the change set is being created...', fg=const.FG_INF)
 
     eResponse = ''
@@ -96,15 +163,43 @@ def _wait_for_done_changeset(name, cName):
     
     while eResponse not in [const.CS_CREATE_COMPLETE, const.CS_FAILED]:
         print(const.ANIM_STRING[anim_index % const.ANIM_LEN], end="\r")   
-        if running_time % const.DELAY_TIME_FOR_DESCRIBE_CHANGE_SET == 0:
+        
+        if running_time % const.DELAY_TIME_FOR_DESCRIBE_CHANGE_SET == 0.0:
             response = _session.cf.describe_change_set(StackName=name,
                                         ChangeSetName=cName) 
+            eResponse = response[const.CS_STATUS]
+
+        anim_index += 1
+        time.sleep(const.DELAY_TIME_FOR_ANIMATION)
+        running_time += const.DELAY_TIME_FOR_ANIMATION
+    return response
+
+def _wait_for_executed_changeset(name):  
+    click.secho('Please wait while the change set is being executed...', fg=const.FG_INF)
+
+    eResponse = ''
+    response = None   
+    running_time = 0
+    anim_index = 0
+    
+    while eResponse not in [const.CREATE_COMPLETE,
+                            const.UPDATE_COMPLETE,
+                            const.DELETE_COMPLETE,
+                            const.UPDATE_ROLLBACK_COMPLETE,
+                            const.ROLLBACK_COMPLETE,
+                            const.IMPORT_ROLLBACK_COMPLETE,
+                            const.IMPORT_COMPLETE]:
+
+        print(const.ANIM_STRING[anim_index % const.ANIM_LEN], end="\r")   
+        
+        if running_time % const.DELAY_TIME_FOR_DESCRIBE_CHANGE_SET == 0.0:
+            response = _session.cf.describe_stacks(StackName=name) 
+            response = response[const.STACKS][0]
+            eResponse = response[const.STACK_STATUS]
    
-        eResponse = response[const.CS_STATUS]
         anim_index += 1
         time.sleep(const.DELAY_TIME_FOR_ANIMATION)
         running_time += const.DELAY_TIME_FOR_ANIMATION
 
-    click.secho('The change set is ready!', fg=const.FG_INF)
-
     return response
+
